refactor(members): start working through register pages; update RegisterStepper
