+++ src/components/shared/HeroBannerImage.tsx
@@ -85,6 +85,7 @@ const StyledRoot = styled.section<HomeBannerImageStyledRootProps>(({
       textAlign: 'center',
     },
     fontSize: 27,
+    fontWeight: 500,
     marginTop: theme.spacing(5),
     textAlign: $leftAlign ? 'left' : 'right',
   },
@@ -156,6 +157,7 @@ const HeroBannerImage: FC<Props> = ({
               colorVariant="about"
               fontWeight={400}
               rightArrow
+              size="large"
               to="/members/join"
               width={224}
             >
